修复问题
  - .github/workflows/ci.yml — GitHub Actions CI（fmt + clippy + build + test + doc）
  - Makefile — make check 一键执行 fmt-check + clippy + test + build
  - Clippy 0 非 trivial 警告（auto-fix：3× map_or → is_none_or/is_some_and，1× 多余 cast）

+++ crates/canal-sink/src/sink.rs
@@ -120,7 +120,7 @@ impl DefaultEventSink {
             while let Some(req) = rx.recv().await {
                 for conn in &connectors {
                     match conn.dispatch(&req.events).await {
-                        Ok(()) => metrics.inc_dispatched(req.filtered_count as u64),
+                        Ok(()) => metrics.inc_dispatched(req.filtered_count),
                         Err(e) => {
                             error!("Connector {} dispatch failed: {}", conn.name(), e);
                             metrics.inc_dispatch_errors(1);
